requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==38.0.1
DateTime==4.7
Django==4.1.2
djangorestframework==3.14.0
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.5.0
PyMySQL==1.0.2
pytz==2022.4
sqlparse==0.4.3
zope.interface==5.5.0
